scripts/mscp/requirements.txt
pydantic>=2.5.0
pyyaml>=6.0.1
typing-extensions>=4.8.0
pytest>=7.4.0